Explicitly close deleted files to avoid leaking
It seems like we're not correctly handling the situation where files are
deleted, and they're kept open for as long as the process is running.

This is probably because we're relying on an implicit `close()`
happening on the file object when the `TailedFile` object gets garbage
collected, but we don't necessarily have control about this (and it
doesn't seem to happen, hinting at other references pointing to it).

This PR adds and explicit `_close()` method to ensure we no longer need
to rely in this implicit behaviour.